(PC-12779)(PRO]feat:sendinblue email validation for new pro user
