+++ api/src/pcapi/admin/custom_views/admin_user_view.py
@@ -10,10 +10,10 @@ from wtforms.validators import Length
 
 from pcapi.admin.base_configuration import BaseAdminView
 from pcapi.admin.custom_views.mixins.suspension_mixin import SuspensionMixin
+from pcapi.core.mails.transactional.pro.email_validation import send_email_validation_to_admin_email
 from pcapi.core.users import api as users_api
 from pcapi.core.users.constants import RESET_PASSWORD_TOKEN_LIFE_TIME_EXTENDED
 from pcapi.core.users.utils import sanitize_email
-from pcapi.domain.user_emails import send_admin_user_validation_email
 from pcapi.utils.mailing import build_pc_webapp_reset_password_link
 
 
@@ -104,7 +104,7 @@ class AdminUserView(SuspensionMixin, BaseAdminView):
             token = users_api.create_reset_password_token(
                 model, token_life_time=RESET_PASSWORD_TOKEN_LIFE_TIME_EXTENDED
             )
-            send_admin_user_validation_email(model, token)
+            send_email_validation_to_admin_email(model, token)
             flash(f"Lien de réinitialisation du mot de passe : {build_pc_webapp_reset_password_link(token.value)}")
 
         super().after_model_change(form, model, is_created)
