+++ api/src/pcapi/domain/user_emails.py
@@ -8,7 +8,6 @@ from pcapi.core.bookings.models import IndividualBooking
 from pcapi.core.offerers.models import Offerer
 from pcapi.core.offerers.repository import find_new_offerer_user_email
 from pcapi.core.users import api as users_api
-from pcapi.core.users.models import Token
 from pcapi.core.users.models import User
 from pcapi.emails import beneficiary_activation
 from pcapi.emails.new_offerer_validated_withdrawal_terms import (
@@ -18,8 +17,6 @@ from pcapi.emails.offerer_booking_recap import retrieve_data_for_offerer_booking
 from pcapi.emails.offerer_bookings_recap_after_deleting_stock import (
     retrieve_offerer_bookings_recap_email_data_after_offerer_cancellation,
 )
-from pcapi.utils.mailing import make_admin_user_validation_email
-from pcapi.utils.mailing import make_pro_user_validation_email
 
 
 logger = logging.getLogger(__name__)
@@ -39,16 +36,6 @@ def send_offerer_bookings_recap_email_after_offerer_cancellation(bookings: list[
         return True
     data = retrieve_offerer_bookings_recap_email_data_after_offerer_cancellation(bookings)
     return mails.send(recipients=[offerer_booking_email], data=data)
-
-
-def send_pro_user_validation_email(user: User) -> bool:
-    data = make_pro_user_validation_email(user)
-    return mails.send(recipients=[user.email], data=data)
-
-
-def send_admin_user_validation_email(user: User, token: Token) -> bool:
-    data = make_admin_user_validation_email(user, token.value)
-    return mails.send(recipients=[user.email], data=data)
 
 
 def send_activation_email(user: User, reset_password_token_life_time: typing.Optional[timedelta] = None) -> bool:
